resovled pmd dependencies
