+++ build.gradle
@@ -4,6 +4,7 @@ plugins {
     id 'io.spring.dependency-management' version '1.1.0'
     id "com.google.protobuf" version "0.9.2"
     id 'checkstyle'
+    id 'pmd'
 }
 
 group = 'com.nucleiassignment3'
@@ -47,6 +48,7 @@ dependencies {
     implementation 'javax.annotation:javax.annotation-api:1.3.2'
     implementation 'io.github.lognet:grpc-spring-boot-starter:5.0.0'
     implementation 'net.devh:grpc-server-spring-boot-starter:2.14.0.RELEASE'
+
 }
 
 checkstyleMain {
@@ -105,6 +107,12 @@ checkstyleMain {
 }
 checkstyleTest {
     source = 'src/test/java'
+}
+pmd {
+    toolVersion = "6.55.0"
+    ignoreFailures = true
+    ruleSetFiles = files("config/pmd/pmd.xml")
+    ruleSets = []                                  // The location of the configuration file.
 }
 
 test {
