--- src/main/java/com/nucleiassignment3/HttpServer/HttpServerApplication.java
@@ -1,13 +1,0 @@
-package com.nucleiassignment3.HttpServer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HttpServerApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(HttpServerApplication.class, args);
-  }
-
-}
