--- src/main/java/com/nucleiassignment3/HttpServer/model/CreateEmployeeRequest.java
@@ -1,15 +1,0 @@
-package com.nucleiassignment3.HttpServer.model;
-
-import java.sql.Date;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class CreateEmployeeRequest {
-  private String name;
-  private Date dob;
-  private String gender;
-}
