--- src/main/java/com/nucleiassignment3/HttpServer/model/UpdateEmployeeRequest.java
@@ -1,15 +1,0 @@
-package com.nucleiassignment3.HttpServer.model;
-
-import java.sql.Date;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UpdateEmployeeRequest {
-  private String name;
-  private Date dob;
-  private String gender;
-}
