+++ src/main/java/com/nucleiassignment3/httpserver/HttpServerApplication.java
@@ -1,0 +1,13 @@
+package com.nucleiassignment3.httpserver;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class HttpServerApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(HttpServerApplication.class, args);
+  }
+
+}
