+++ src/main/java/com/nucleiassignment3/httpserver/bo/EmployeeBo.java
@@ -1,0 +1,20 @@
+package com.nucleiassignment3.httpserver.bo;
+
+import java.sql.Date;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class EmployeeBo {
+
+  private String empId;
+  private String name;
+  private Date dob;
+  private String gender;
+
+}
