+++ src/main/java/com/nucleiassignment3/httpserver/model/CreateEmployeeRequest.java
@@ -1,0 +1,15 @@
+package com.nucleiassignment3.httpserver.model;
+
+import java.sql.Date;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class CreateEmployeeRequest {
+  private String name;
+  private Date dob;
+  private String gender;
+}
